Move IntegrationCredentials to testing package, add explicit catch fail to IntegrationCredentialsTest, add an ENV variable for the VAULT_PATH to the service account.
