+++ .github/workflows/build-and-test.yaml
@@ -10,6 +10,7 @@ on:
     branches: [ '**' ]
 env:
   VAULT_ADDR: https://clotho.broadinstitute.org:8200
+  VAULT_SERVIC_ACCOUNT_PATH: secret/dsde/terra/crl-test/default/service-account.json
   # Where to store the retrieved service account credentials for Google integration tests.
   SERVICE_ACCOUNT_FILE: src/test/resources/integration_service_account.json
 jobs:
@@ -38,7 +39,7 @@ jobs:
             -e "VAULT_TOKEN=${{ steps.vault-token-step.outputs.vault-token }}" \
             -e "VAULT_ADDR=${VAULT_ADDR}" \
             vault:1.1.0 \
-            vault read -format json secret/dsde/terra/crl-test/default/service-account.json \
+            vault read -format json $VAULT_SERVICE_ACCOUNT_PATH \
             | jq .data > $SERVICE_ACCOUNT_FILE
       - name: Set up JDK 1.8
         uses: actions/setup-java@v1
